Debug geo/time dataframe creation code
Touch up and debug code so that it works across multiple
environments.

Previously, this code had been tested and debugged on
my local machine, but not up on an AWS machine. Throwing
it up to an AWS machine exposed a coule of issues that
helped me re-write and touch up the code to be a little
bit more generalizable.