Edit query for count of nearby fires.
Technically, in real-time we won't have a count of the nearby fires,
since forest-fire perimeter boundaries are submitted at the end of
every day. This means that the previous version of the query,
which acted as if we knew which 'detected fires' were actually
forest-fires for all detected fires with an earlier timestamp
than the row we were querying for (regardless of whether they
were the same day or prior days), was not realistic.