Update Makefile to ouput data
I wanted to add additional step to the Makefile to output all of the
detected fires data into actual CSVS, rather than that having to be a
separate command to run outside of the Makefile. Anything that needs to
happen inside of PostgresSQL will happen before this output, so I don't
see any reason not to add this step into the Makefile.