Update get_train_test with 1 day test, 14 day train
I'm beginning on working to moving the model training/validation to be
more in line with what would be used in a production setting. In a
production setting, we would theoretically train a model each night on
the most relevant data, and then test on observations during the day.

To create as close of an environment to this as I can get, I am using a
test set that is 1 day, and then using the prior 14 days as the training
set (where I will do CV over the training set). Note that these 14 days
won't be what I actually use to train the model (that'll be a different
subset of data that I'll describe later), but just the 14 days that I
will perform CV over (i.e. I'll use each one of the 14 days to test on,
training via the aforementioned subset of data that I'll describe
later).