Rework classes for sequential time fold validation
In moving this towards the production setting, I am rethinking how I am
going to validate, which calls for a reorganization of how my time val
classes will be constructed. Rather than just taking in an initial date,
training with all points that are before and testing on all after (or
something like this), and then adding some interval to the initial date
and doing that again, I'm doing something else.

I'm now going to pass in one date (the lastest day in the train
set). This date will be the first testing date (used for validation),
and then we'll choose what we train on either sequentially (all prior
obs.) or stratified (all those in the same weeks of prior years). We'll
do this for the first testing date, then go backwards by the step size,
and do this over again. We'll do this until we hit the number of folds
that is passed in.

Any of the changes in run_model.py in this commit were to get this
change working (for example now we simply have test that is 1 day,
and train as all the remaining days (instead of 14 like a prior
commit mentioned)).