Add alt train set code if pct fires is too small
If the percent of fires in the training set is too small, then I want to
do something about that. There are two scenarios that I'm going to
tackle. The first is it being absurdly small (like < 5%, or < 1%), to
the point that even upsampling/downsampling would be unlikely to help.
The second is it being in some mid-range, where upsampling/downsampling
could help (5-30%, or something like that).

Here I've coded up the case that is absurdly small. If that's the case,
then we are with near certainty using the stratified code (if we were
using sequential, then we would be grabbing all previous obs., and the
average % fires there is around 13-17). In this case, a simple solution
is to move to using the sequential code for that fold (i.e. grab all
prior obs., not just those that are in some number of days prior for
each prior year).