Clean up and update eda-viz folder
The eda-viz folder was a little bit of a mess, and if I were to look at
it in a year, I probably wouldn't have a good idea of how I worked
through things. In this commit I deleted things that are no longer
useful (or weren't worth saving in the first place), as well as cleaned
up and reorganized what was left.