Update Makefile to differentiate MODIS data
I'm switching gears to modify the data gathering process (I'll be on a
plane in a couple of days, where I can continue the viz. stuff without
the Internet). I need to start working on integrating the VIIRS data, as
it is finer resolution. I'll be starting with the VIIRS 750m data and
seeing if/how that helps. From there, I might move to looking at the
VIIRS I Band (350m) data. This commit is the first step in doing that -
I need to differentiate the data gathering process based off what source
the data is coming from.