Update dist exploration for new data and code
I've reworked the repo. so that the data download step will also include
grabbing the VIIRS data, in addition to MODIS. As such, I needed to
change the filepath of the data that the dist. exploration notebook was
reading in. I also made some slight alterations to some of the plotting
functionality that the notebook was using, and needed to update it to
account for those changes.