Add functionality to grab/plot perimeter boundaries
With this commit, I've added one main function, which will allow
plotting of a county and all it's observations and fire perimeter
boundaries for an inputted state. Along the way, I touched up/moved a
few functions (in the name of the DRY methodology) as well as built a
bash script to grab a particular subset of fire perimeter boundaries out
of postgres for me.