Add functionality for plotting on a single day
One of the only pieces of functionality that I hadn't built up to this
point was the functionality to plot a county, and it's fire perimeters
and fires for a single day. It's now here.

I've also ported over the use of the StateMap and CountyMap classes to a
USMapBuilder class. This accompanies a change in the API for the dsfuncs
module used during EDA.