Adjust validation strategy
I'm choosing to validate over a month of data at a time, going through
2014 and using 2015 as the hold out set (to be tested on after selection
a model specification). I've adjusted run_model.py and time_val.py to
take account for this change in validation strategy. Basically, trying
to validate on a single day earlier was too noisy, even if predictions
in near real-time will only be on a single day (or less). Validating on
a month at a time (rather than a single day) should help to actually
select the best model specification.